# apps/frontend/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_decode_tac_endpoint_api_v1_decode_tac_post": {
        "properties": {
          "files": {
            "anyOf": [
              {
                "items": {
                  "contentMediaType": "application/octet-stream",
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Files"
          },
          "manual_text": {
            "default": "",
            "description": "TAC text to decode",
            "title": "Manual Text",
            "type": "string"
          },
          "product": {
            "description": "TAC product (required)",
            "title": "Product",
            "type": "string"
          }
        },
        "required": [
          "product"
        ],
        "title": "Body_decode_tac_endpoint_api_v1_decode_tac_post",
        "type": "object"
      },
...
      "Body_lint_tac_api_v1_lint_tac_post": {
        "properties": {
          "files": {
            "anyOf": [
              {
                "items": {
                  "contentMediaType": "application/octet-stream",
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Files"
          },
          "manual_text": {
            "default": "",
            "description": "TAC or IWXXM XML to lint",
            "title": "Manual Text",
            "type": "string"
          },
          "product": {
            "default": "METAR",
            "description": "Product type, or iwxxm for XML lint (default METAR)",
            "title": "Product",
            "type": "string"
          }
        },
        "title": "Body_lint_tac_api_v1_lint_tac_post",
        "type": "object"
      },
...
      "DecodeResidualModel": {
        "description": "HTTP DTO for an undecoded TAC span (explicit residuals - G4).",
        "properties": {
          "end": {
            "title": "End",
            "type": "integer"
          },
          "start": {
            "title": "Start",
            "type": "integer"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "start",
          "end",
          "text"
        ],
        "title": "DecodeResidualModel",
        "type": "object"
      },
      "DecodeSegmentModel": {
        "description": "HTTP DTO for one TAC decode/annotate segment.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "end": {
            "title": "End",
            "type": "integer"
          },
          "explanation": {
            "title": "Explanation",
            "type": "string"
          },
          "start": {
            "title": "Start",
            "type": "integer"
          }
        },
        "required": [
          "start",
          "end",
          "code",
          "explanation"
        ],
        "title": "DecodeSegmentModel",
        "type": "object"
      },
      "DecodeTacResponse": {
        "description": "Response for POST /api/v1/decode-tac.",
        "properties": {
          "product": {
            "title": "Product",
            "type": "string"
          },
          "residuals": {
            "items": {
              "$ref": "#/components/schemas/DecodeResidualModel"
            },
            "title": "Residuals",
            "type": "array"
          },
          "segments": {
            "items": {
              "$ref": "#/components/schemas/DecodeSegmentModel"
            },
            "title": "Segments",
            "type": "array"
          },
          "summary": {
            "default": "",
            "description": "Deterministic plain-language paragraph of the report",
            "title": "Summary",
            "type": "string"
          }
        },
        "required": [
          "product"
        ],
        "title": "DecodeTacResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LintFixModel": {
        "description": "HTTP DTO for an optional tac-validate fix suggestion.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "replacement": {
            "title": "Replacement",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "replacement"
        ],
        "title": "LintFixModel",
        "type": "object"
      },
      "LintIssueCatalogEntryModel": {
        "description": "One registry row exported by GET /api/v1/lint-issue-catalog.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "exchange_profiles": {
            "description": "Canonical exchange profile ids for packaging-tagged rows; empty = shared",
            "items": {
              "type": "string"
            },
            "title": "Exchange Profiles",
            "type": "array"
          },
          "family": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "lint (TAC registry) or iwxxm (validation checks)",
            "title": "Family"
          },
          "issue_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Closed vocabulary: presence, structure, content, consistency, iwxxm_schema, other",
            "title": "Issue Type"
          },
          "last_verified": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "ISO date of last HTTP check for operator source_url",
            "title": "Last Verified"
          },
          "message_template": {
            "title": "Message Template",
            "type": "string"
          },
          "product": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Product"
          },
          "replacement_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Verified landing when source_url is a legacy alias",
            "title": "Replacement Url"
          },
          "semantic_identifier": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Vocabulary concept path when href is a verified landing",
            "title": "Semantic Identifier"
          },
          "semantic_profiles": {
            "description": "Canonical semantic profile ids this row applies to; empty = shared/global",
            "items": {
              "type": "string"
            },
            "title": "Semantic Profiles",
            "type": "array"
          },
          "severity": {
            "title": "Severity",
            "type": "string"
          },
          "source_access": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Operator access tier: public, paywall, login, semantic_only",
            "title": "Source Access"
          },
          "source_attribution": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Attribution"
          },
          "source_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Id"
          },
          "source_locator": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Section/table/page locator for the cited source",
            "title": "Source Locator"
          },
          "source_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "tier1, tier2, or tier3 source policy",
            "title": "Source Type"
          },
          "source_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "verified, legacy_alias, or semantic_only",
            "title": "Status"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          }
        },
        "required": [
          "code",
          "severity",
          "message_template"
        ],
        "title": "LintIssueCatalogEntryModel",
        "type": "object"
      },
      "LintIssueCatalogResponse": {
        "description": "Response for GET /api/v1/lint-issue-catalog.",
        "properties": {
          "issues": {
            "items": {
              "$ref": "#/components/schemas/LintIssueCatalogEntryModel"
            },
            "title": "Issues",
            "type": "array"
          }
        },
        "title": "LintIssueCatalogResponse",
        "type": "object"
      },
      "LintIssueModel": {
        "description": "HTTP DTO for a tac-validate issue (msgspec \u2192 pydantic).",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "end": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Exclusive character offset",
            "title": "End"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "severity": {
            "title": "Severity",
            "type": "string"
          },
          "start": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Inclusive character offset",
            "title": "Start"
          }
        },
        "required": [
          "severity",
          "code",
          "message"
        ],
        "title": "LintIssueModel",
        "type": "object"
      },
      "LintTacResponse": {
        "description": "Response for POST /api/v1/lint-tac.",
        "properties": {
          "fixes": {
            "items": {
              "$ref": "#/components/schemas/LintFixModel"
            },
            "title": "Fixes",
            "type": "array"
          },
          "issues": {
            "items": {
              "$ref": "#/components/schemas/LintIssueModel"
            },
            "title": "Issues",
            "type": "array"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          },
          "product": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Product"
          }
        },
        "required": [
          "ok"
        ],
        "title": "LintTacResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Convert METAR/SPECI TAC messages to IWXXM XML format with comprehensive validation",
    "title": "METAR to IWXXM Backend API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/decode-tac": {
      "post": {
        "description": "Decode TAC into annotated segments and a plain-language summary.",
        "operationId": "decode_tac_endpoint_api_v1_decode_tac_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_decode_tac_endpoint_api_v1_decode_tac_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecodeTacResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "415": {
            "description": "Unsupported Media Type - multipart/form-data required"
          },
          "422": {
            "description": "Missing required product field"
          }
        },
        "summary": "Decode Tac Endpoint",
        "tags": [
          "Validation",
          "Conversion"
        ]
      }
    },
...
    "/api/v1/lint-issue-catalog": {
      "get": {
        "description": "Export TAC lint + IWXXM validation catalog for FE tooltips / catalog page.",
        "operationId": "lint_issue_catalog_api_v1_lint_issue_catalog_get",
        "parameters": [
          {
            "in": "query",
            "name": "product",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Product"
            }
          },
          {
            "in": "query",
            "name": "family",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Family"
            }
          },
          {
            "in": "query",
            "name": "issue_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Issue Type"
            }
          },
          {
            "in": "query",
            "name": "source_access",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source Access"
            }
          },
          {
            "in": "query",
            "name": "semantic_profile",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Semantic Profile"
            }
          },
          {
            "in": "query",
            "name": "exchange_profile",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Exchange Profile"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LintIssueCatalogResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Lint Issue Catalog",
        "tags": [
          "Validation"
        ]
      }
    },
    "/api/v1/lint-tac": {
      "post": {
        "description": "Thin wrapper over ``packages/tac-validate`` (multipart/form-data only - Q8=A).",
        "operationId": "lint_tac_api_v1_lint_tac_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_lint_tac_api_v1_lint_tac_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LintTacResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "415": {
            "description": "Unsupported Media Type - multipart/form-data required"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Lint Tac",
        "tags": [
          "Validation"
        ]
      }
    },
...
  },
...
}